requirements.txt
asgiref
python-decouple
gunicorn
django-mptt
setuptools
dj-database-url
cachetools
whitenoise
certifi
charset-normalizer
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
protobuf
pyasn1
pyasn1-modules
PyJWT
pyparsing
pytz
requests
rsa
six
sqlparse
psycopg2
tzdata
uritemplate
urllib3

tensorflow
numpy
pandas
scikit-learn
geopy